Diagnostic plot

Generates diagnostic curves for evaluation of calibration. Require that data have been calibrated.

diagnosticPlot(
  object,
  which.assay = NULL,
  which.parameter = NULL,
  which.phantom = NULL,
  which.plot = "bar",
  highlight.site = NULL,
  color.option = "cividis",
  color.begin = 0,
  color.end = 1,
  bar.errors = F,
  fix.axis = T,
  return.plt.handle = F
)

Arguments

object
Calibration Object

which.assay
Character specifying which assay to use.

which.parameter
Character indicating which parameters to plot diagnostics for. If unspecified, all parameters are plotted.

which.phantom
Character specifying which phantom to use.

which.plot
Character indicating which diagnostic plot to produce. One of:

"line" - Line plot used to compare pre and post calibrated values.

"bar" - Horizontal barplots used to compare pre and post calibrated values

"residual" - Histogram used to visualize pre- and post-calibration residuals (i.e., difference between overall mean and replicate mean)

"dif" - Histogram used to visualise percentage difference between pre and post-calibrated values.

highlight.site
Character specifying site to highlight in diagnostic plot (valid for line plots)

color.option
Character indicating which color map option to use (from viridis palette). One of:

"magma" (or "A")

"inferno" (or "B")

"plasma" (or "C")

"viridis" (or "D")

"cividis" (or "E") - Default

color.begin
Hue in [0,1] at which the viridis color map begins

color.end
Hue in [0,1] at which the viridis color map ends

bar.errors
Logical indicating whether to show error bars for barplot (valid for bar plots)

fix.axis
Logical indicating whether to fix x-axis. Default is True. (Valid for bar and residual plots)

return.plt.handle
Logical specifying whether to return list of plot handles. If TRUE, list of plot handles is returned. If FALSE, plots are printed without returning handle.

Value

plot handle (see return.plt.handle argument)

Details

Diagnostics are only run for data originating from the calibrating phantom. I.e.,if data from multiple phantoms are present within the dataset, only the data from the calibrating phantom is retained for diagnostic purposes.

See also