Fit calibration curves

Fit pairwise linear regression between calibration and reference sites, using uncalibrated dataset in specified Assay.

fitCalibration(
  object,
  reference.site = NULL,
  reference.time = "baseline",
  which.phantom = NULL,
  sig.intercept.only = F,
  which.parameter = NULL,
  omit.parameter = NULL,
  which.assay = NULL,
  n.signif = 3,
  verbose = T,
  which.center = "mean"
)

Arguments

object
Calibration Object

reference.site
Character specifying reference site (i.e., reference instrument to calibrated other instruments)

reference.time
Character specifying reference time. Does not need to be specified if only one timePoint is available. One of:

baseline - Default. All calibration curves are generated using reference site measurements at baseline as the reference timpoint.

match - Calibration curves are constructed using matched time points. E.g., Site measurements are baseline are calibrated using reference site measurements at baseline, 6 month measurements are calibrated using references measurements at 6 months, etc...

which.phantom
Character indicating which phantom to fit calibration curves for. If unspecified, assumed that only one phantom is provided in data. In this case, if multiple exist, and error will occur.

sig.intercept.only
Logical specifying whether to include intercept in fitted regression equation.

TRUE - Only include intercept in model if significant (p < 0.05).

FALSE - Default. Include intercept in model regradless of significance.

which.parameter
Character specifying which parameters to fit calibration curves for. If unspecified, all parameters are fit.

omit.parameter
Character specifying which parameters to omit from calibration curve fitting. Overrides parameters specified by which.parameter argument.

which.assay
Character specifying which assay to fit calibration curves for.

n.signif
Number of significant figures to report.

verbose
Logical specify whether to report progress.

which.center
Specify which measure of central tendency to use when fitting calibration curves. One of:

"mean" - Deafult. mean replicate values.

"median" - median replicate values. Recommended if >2 replicates scanned, and outliers are present.

Value

Calibration Object

See also