Multi-variant reproducibility boxplots

Generates boxplots summarizing results generated during mvp.analysis.

mvpPlot(
  object,
  which.data = "all",
  group.by = NULL,
  which.parameter = NULL,
  which.phantom = NULL,
  which.precision = "all",
  which.analysis = NULL,
  var2plot = "cv",
  which.assay = NULL,
  outliers = T,
  jitter.width = 0.1,
  color.option = "cividis",
  color.begin = 0,
  color.end = 1,
  point.size = 2,
  point.alpha = 1,
  box.alpha = 0.5,
  return.plt.handle = F,
  combine.plots = F,
  combine.ncols = 2,
  show.rms.statistic = F,
  show.legend.flag = T
)

Arguments

object
Calibration Object.

which.data
Character specifying which dataset to plot svp analysis for. One of:

"all" - Default

"uncalibrated"

"calibrated"

group.by
Vector of features to stratify analysis by. If specified, rms-statistic is not overlaid on boxplot.

which.parameter
Character specifying which parameters to plot. If unspecified, all parameters are plotted.

which.phantom
Character specifying which phantoms to plot. If unspecified, all phantoms are plotted.

which.precision
Character specifying which precision errors to plot. One of:

"all" - Default. Short and longitudinal, single- and multi-site precision errors

"short" - Short-term precision errors only

"long" - Longitudinal precision errors only

"single" - Single-site precision errors only

"multi" - Multi-site precision errors only

which.analysis
Character indicating which mvp.analysis to get results from. If unspecified, available analyses will be used.

var2plot
Character specifying which precision error metric to plot. One of:

cv - coefficient of variance

std - standard deviation

which.assay
Character specifying which assay to plot.

outliers
Logical specifying whether to omit outliers from analysis

jitter.width
Numerical specifying width of jitter in boxplot. Default is 0.1.

color.option
Character indicating which color map option to use (from viridis palette). One of:

"magma" (or "A")

"inferno" (or "B")

"plasma" (or "C")

"viridis" (or "D")

"cividis" (or "E") - Default

color.begin
Hue in [0,1] at which the viridis color map begins. Default is 0.

color.end
Hue in [0,1] at which the viridis color map ends. Default is 1.

point.size
Numeric specifying size of scatter plot points.

point.alpha
Numeric [0,1] specifying degree of transparency for points. Default is 1.

box.alpha
Numeric [0,1] specifying degree of transparency for box plots. Default is 0.5.

return.plt.handle
Logical indicating whether list of ggplot handle is returned. Default is false.

combine.plots
Logical indicating whether plots are combined. Default is false.

combine.ncols
Number of columns when combining plots (Only if combine.plot is true). Default is 2.

show.rms.statistic
Logical indicating to overlay RMS statistic (rms-std or rms-cv) on boxplots. Default is false.

show.legend.flag
Logical indicating to show legend

Value

list of ggplot objects (if return.plt.handle is true)

See also