Single variant precision/reproducibility analysis

Short-term single-center precision analysis of repeated phantom/in vivo measurements. Parameter-specific precision errors are calculated as root-mean-square standard deviation (rms.std) and root-mean-square coefficient of variance (rms.cv). Results are saved in calibration object as 'svp.analysis'.

svpAnalysis(
  object,
  which.data = "uncalibrated",
  replicate.strata = NULL,
  n.signif = 3,
  which.assay = NULL,
  verbose = T
)

Arguments

object
Calibration Object

which.data
A character specifying which data to analyze. One of:

uncalibrated (Default)

calibrated

replicate.strata
Vector specifying how a set of scan replicates are defined. If specified, overrides replicateSet definition. If replicateSet definition not present in Calibration Object and replicate.strata is not specified, default replicate definition is stratification by site, timePoint, section, parameter, phantom. That is, all scans grouped by the specified features, and pooled within each strata prior to pooling across strata to compute parameter-specific rms-statistics.

n.signif
Number of significant digits reported.

which.assay
A character specifying assay for analysis. If unspecified, set to current assay.

verbose
Logical specifying where to report progress.

Value

Calibration Object

See also